requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==4.1.0
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
cssselect2==0.4.1
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==3.0.1
django-bower==5.2.0
django-ckeditor==6.1.0
django-contrib-comments==2.1.0
django-filer==2.1.1
django-js-asset==1.2.2
django-libs==2.0.3
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications-hq==1.6.0
django-password-reset==2.0
django-polymorphic==3.0.0
django-private-chat2==0.1.4
django-scheduler==0.9.5
django-swingtime==1.2.1
django-tagging==0.5.0
django-taggit==1.5.1
django-webpush==0.3.3
django-widget-tweaks==1.4.9
django-xmlrpc==0.1.8
docker==4.3.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.18.1
easy-thumbnails==2.8
filelock==3.4.0
gitdb==4.0.9
GitPython==3.1.7
google-auth==2.3.3
gunicorn==20.1.0
hiredis==2.0.0
http-ece==1.1.0
hyperlink==21.0.0
icalendar==4.0.9
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
invoke==1.6.0
jsonfield==3.1.0
keyring==23.2.1
kombu==5.2.2
kubernetes==11.0.0
lxml==4.6.4
Markdown==3.3.4
mock==4.0.3
mots-vides==2015.5.11
msgpack==1.0.2
nifty==1.6.3
oauthlib==3.1.1
packaging==21.3
parsel==1.6.0
Pillow==8.4.0
pkginfo==1.8.1
platformdirs==2.4.0
pluggy==1.0.0
popper==2020.9.1
prompt-toolkit==3.0.24
py==1.11.0
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pykwalify==1.7.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
python-box==5.1.1
python-dateutil==2.8.0
pytz==2021.3
pywebpush==1.9.4
PyYAML==5.3.1
readme-renderer==30.0
redis==4.0.1
regex==2021.11.2
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
semver==2.13.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
soupsieve==2.2.1
spython==0.0.79
sqlparse==0.4.2
static3==0.7.0
svglib==1.1.0
swapper==1.3.0
testify==0.11.0
tinycss2==1.1.0
toml==0.10.2
tox==3.24.4
tqdm==4.62.3
twine==3.6.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.0
Unidecode==1.1.2
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
wrapt==1.13.3
zest.releaser==6.22.2
zipp==3.6.0
zope.interface==5.4.0
